requirements.txt
click
pytest
solidpython==0.4.4
